Change to v2_routeConvention, moving and renaming some resource routes accordingly
