+++ app/routes/resource.get-blog-post.$slug.ts
