+++ remix.config.js
@@ -7,6 +7,7 @@ module.exports = {
     v2_errorBoundary: true,
     v2_meta: false,
     v2_normalizeFormMethod: true,
+    v2_routeConvention: true,
   },
   ignoredRouteFiles: ["**/.*", "**/*.css", "**/*.test.{js,jsx,ts,tsx}"],
   serverDependenciesToBundle: [
